refactor: WishlistRepository의 PK 타입 Long으로 수정

+++ src/main/java/gift/wishlist/repository/WishlistRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WishlistRepository extends JpaRepository<Wishlist, Integer> {
+public interface WishlistRepository extends JpaRepository<Wishlist, Long> {
 
     Optional<Wishlist> findByMemberIdAndProductId(Long memberId, Long productId);
